Update link path
I already updated the Wiki by hand, sorry for not reading the text at the top first. 

Somebody changed the path structure of the project, and that broken all the links to the Haxe and Python sources. 

Not only this, but also other parts of the documentation might be damaged.